Move TinyVector to a new file
- So it's less sensitive to #include ordering
